+++ desktop/Rectangle.cpp
@@ -18,7 +18,7 @@ Rectangle::Rectangle(const Rectangle& rectangle)
 }
 
 
-Rectangle::Rectangle(std::string name, int position[], float velocity[], int colour[], float scale[], int width, int height)
+Rectangle::Rectangle(std::string name, int position[], float velocity[], float colour[], float scale[], int width, int height)
 	: Shape(name, position, velocity, colour, scale)
 	, m_width(width)
 	, m_height(height)
